% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_tree.R
\name{add_tree}
\alias{add_tree}
\title{Attach phylogenetic tree to the LorMe object}
\usage{
add_tree(taxobj, tree)
}
\arguments{
\item{taxobj}{One LorMe object. Should be generated from \code{\link{tax_summary}}}

\item{tree}{The phylogenetic tree. Recommanded to load by \code{\link[ape]{read.tree}} or \code{\link[phyloseq]{phy_tree}}}
}
\value{
A LorMe object with phylogenetic tree
}
\description{
Attach phylogenetic tree to the LorMe object
}
\examples{
data("Two_group")
}
