% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Alpha_diversity_calculator.R
\name{Alpha_diversity_calculator}
\alias{Alpha_diversity_calculator}
\title{Calculate alpha diversity based on tax summary object}
\usage{
Alpha_diversity_calculator(taxobj, taxlevel, prefix = "")
}
\arguments{
\item{taxobj}{tax summary objects computed by \code{\link{tax_summary}}. Default:NULL.}

\item{taxlevel}{taxonomy levels used for visualization.Must be one of c("Domain","Phylum","Class","Order","Family","Genus","Species","Base").Default:NULL.}

\item{prefix}{A character string as prefix of diversity index. Default:""}
}
\value{
'Alpha_diversity_calculator' returns α-diversity of each sample in format of column table (dataframe) combined with group information in meta file.
}
\description{
Calculate alpha diversity for each sample
}
\examples{
###data prepration####
data("Two_group")

###analysis####
Alpha<- Alpha_diversity_calculator(taxobj = Two_group,taxlevel = "Base")
head(Alpha)

Alpha<- Alpha_diversity_calculator(taxobj = Two_group,taxlevel = "Genus")
}
