% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.LorMe_define.R
\name{$,LorMe-method}
\alias{$,LorMe-method}
\title{Access LorMe object elements via $}
\usage{
\S4method{$}{LorMe}(x, name)
}
\arguments{
\item{x}{A LorMe object}

\item{name}{Character: Slot name or key in the 'data' slot}
}
\value{
Value of the specified slot or data element
}
\description{
S4 method to enable $ operator for LorMe objects, supporting both slots and
elements in the 'data' slot.
}
