% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.LorMe_define.R
\name{.DollarNames.LorMe}
\alias{.DollarNames.LorMe}
\title{Tab completion for LorMe $ access}
\usage{
\method{.DollarNames}{LorMe}(x, pattern = "")
}
\arguments{
\item{x}{A LorMe object}

\item{pattern}{Character: Prefix pattern for matching}
}
\value{
Character vector of valid keys matching the pattern
}
\description{
enable tab completion when using $ with LorMe objects, including
both slot names and keys in the 'data' slot.
}
