% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_options.R
\docType{data}
\name{.LorMe_opts}
\alias{.LorMe_opts}
\title{Global options for LorMe pipeline analysis}
\format{
An object of class \code{list} of length 11.
}
\usage{
.LorMe_opts
}
\description{
Global options for LorMe pipeline analysis
}
\examples{
getOption("LorMe")
}
\keyword{datasets}
