% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Facet_group.R
\docType{data}
\name{Facet_group}
\alias{Facet_group}
\title{Tax summary object with Facet 2x2 Groups}
\format{
Tax summary object with configuration
}
\usage{
Facet_group
}
\description{
Enraptured summary object with facet 2x2 Groups.Configuration has been assigned.
}
\keyword{datasets}
