% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LorMe_defaults.R
\name{LorMe_defaults}
\alias{LorMe_defaults}
\title{Restore LorMe global options to factory defaults}
\usage{
LorMe_defaults()
}
\value{
Invisible NULL
}
\description{
Equivalent to \code{options(LorMe = .LorMe_opts)} but provides a unified entry.
}
\examples{
LorMe_defaults()   # one-click factory reset
}
