% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LorMe_options.R
\name{LorMe_options}
\alias{LorMe_options}
\title{Change LorMe options}
\usage{
LorMe_options(..., level = NULL)
}
\arguments{
\item{...}{Named parameters or "level = list" pairs.}

\item{level}{Optional; specifies a whole-layer replacement.}
}
\value{
Modified global options (or invisible(NULL) after factory reset).
}
\description{
Change LorMe options
}
\examples{
LorMe_options()                                    # view current
LorMe_options(beta = list(diagram = "circle"))     # whole layer
LorMe_options(level = "beta", list(diagram = "circle"))  # within layer
}
