% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_stat.R
\name{network_stat}
\alias{network_stat}
\title{Igraph network statistics}
\usage{
network_stat(input, report = TRUE)
}
\arguments{
\item{input}{An igraph object.}

\item{report}{Logical. If print report to console. Default:TRUE}
}
\value{
network statistics
}
\description{
Igraph network statistics
}
\author{
Wang Ningqi \href{mailto:2434066068@qq.com}{2434066068@qq.com}
}
