% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_run.R
\name{safe_run}
\alias{safe_run}
\title{Running function with safe mode}
\usage{
safe_run(func)
}
\arguments{
\item{func}{Functions to run}
}
\value{
Same as of functions
}
\description{
Running function with safe mode
}
\examples{
safe_run(print("Hello"))
}
