% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.LorMe_define.R
\name{show,LorMe-method}
\alias{show,LorMe-method}
\title{Show method for LorMe objects}
\usage{
\S4method{show}{LorMe}(object)
}
\arguments{
\item{object}{A LorMe object}
}
\description{
Custom summary display when LorMe object is printed.
}
