% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testotu.R
\docType{data}
\name{testotu}
\alias{testotu}
\title{test otudata}
\format{
A data frame with 1000 rows and 22 variables.
}
\usage{
testotu
}
\description{
A dataset containing 20 samples and 1000 OTUs from soil to test,taxonomy information
is covered randomly(not actual)
}
\keyword{datasets}
