% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_zg.R
\name{theme_zg}
\alias{theme_zg}
\title{A classic theme for ggplot}
\usage{
theme_zg()
}
\value{
ggplot theme
}
\description{
A classic theme for ggplot
}
\note{
Build inside the LorMe package, Please use theme_zg() as a theme directly
}
