% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Three_group.R
\docType{data}
\name{Three_group}
\alias{Three_group}
\title{Tax summary object with three groups}
\format{
Tax summary object with configuration
}
\usage{
Three_group
}
\description{
Enraptured summary object with three groups.Configuration has been assigned.
}
\keyword{datasets}
