% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Two_group.R
\docType{data}
\name{Two_group}
\alias{Two_group}
\title{Tax summary object with two groups}
\format{
Tax summary object with configuration
}
\usage{
Two_group
}
\description{
Enraptured summary object with two groups.Configuration has been assigned.
}
\keyword{datasets}
